Extract XCCDFEntity class to be used in other module files
